Move the edit tests to the store suite
